refactor(test): split oversized test files to respect 300-line limit
